Move `Behaviour` in dedicated module
